Create new file/module inputs for all config reading function from both simulation and scenarios.
