+++ popsborder/inputs.py
@@ -1,0 +1,496 @@
+# Simulation of contaminated consignments and their inspections
+# Copyright (C) 2018-2021 Vaclav Petras and others (see below)
+
+# This program is free software; you can redistribute it and/or modify it under
+# the terms of the GNU General Public License as published by the Free Software
+# Foundation; either version 2 of the License, or (at your option) any later
+# version.
+
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+# details.
+
+# You should have received a copy of the GNU General Public License along with
+# this program; if not, see https://www.gnu.org/licenses/gpl-2.0.html
+
+
+"""Inputs, especially loading of configuration
+
+.. codeauthor:: Vaclav Petras <wenzeslaus gmail com>
+"""
+
+import copy
+import math
+import json
+import sys
+import types
+from collections.abc import Iterable, Mapping
+from pathlib import Path
+
+
+def text_to_value(arg):
+    """Convert text to int, float, or interpret it as JSON if possible
+
+    If the argument is not string, it is returned as is.
+    If conversion is not possible, the original text is returned.
+    None is returned for both None and an empty string.
+    """
+    if not isinstance(arg, str):
+        return arg
+    if not arg:
+        return None
+    try:
+        return int(arg)
+    except ValueError:
+        try:
+            return float(arg)
+        except ValueError:
+            try:
+                return json.loads(arg)
+            except json.JSONDecodeError:
+                # add yaml.safe_load here?
+                # Return the original value.
+                return arg
+
+
+def update_nested_dict_by_dict(dictionary, update):
+    """Recursively update nested dictionary by anther nested dictionary"""
+    for key, value in update.items():
+        if isinstance(value, Mapping):
+            update_nested_dict_by_dict(dictionary.get(key, {}), value)
+        else:
+            dictionary[key] = value
+
+
+def update_nested_dict_by_item(dictionary, keys, value):
+    """Update nested dictionary by a nested keys-value pair
+
+    An item is a list of keys to navigate the nested dictionary and a value to place
+    in the given position.
+
+    When a key can be represented as an int, it is used as a list index.
+    List must already exists in the given size or the only index used must be 0.
+
+    Floating point keys are not supported.
+    """
+    if len(keys) == 1:
+        key = keys[0]
+        try:
+            key = int(key)
+        except ValueError:
+            pass
+        dictionary[key] = value
+    else:
+        if keys[0] not in dictionary:
+            try:
+                # Test if the next key is an integer and thus index in a list.
+                int(keys[1])
+                # In case it is a list, we require keys are items are filled in order.
+                dictionary[keys[0]] = [None]
+            except ValueError:
+                dictionary[keys[0]] = {}
+        update_nested_dict_by_item(dictionary[keys[0]], keys[1:], value)
+
+
+def record_to_nested_dictionary(record):
+    """Convert dictionary with key/subkey/subsubkey keys into a nested dictionary"""
+    out = {}
+    for path, value in record.items():
+        try:
+            keys = path.split("/")
+        except AttributeError as error:
+            raise ValueError(
+                f"Record items need to be strings, not {path} ({type(path).__name__})"
+            ) from error
+        update_nested_dict_by_item(out, keys, value)
+    return out
+
+
+def update_config(config, record):
+    """Update config dictionary by a dictionary with key/subkey/subsubkey keys"""
+    config = copy.deepcopy(config)
+    update = record_to_nested_dictionary(record)
+    update_nested_dict_by_dict(config, update)
+    return config
+
+
+def load_configuration_yaml_from_text(text):
+    """Return configuration dictionary from YAML in a string"""
+    import yaml  # pylint: disable=import-outside-toplevel
+
+    if hasattr(yaml, "full_load"):
+        return yaml.full_load(text)
+    return yaml.load(text)  # pylint: disable=no-value-for-parameter
+
+
+def load_configuration(filename, sheet=None, key_column=None, value_column=None):
+    """Get the configuration from a JSON or YAML file
+
+    The format is decided based on the file extension.
+    It uses full_load() (FullLoader) to read YAML.
+
+    The parameter can be a string or a path object (path-like object).
+
+    If the *filename* contains `::`, anything after the last `::` is considered
+    parameters determining where in the spreadsheet or table are the relevant
+    columns. The format is multiple key-value pairs with key and value separated by
+    `=`, `:`, or `: ` and individual keys separated by `,`.
+    The same information can be passed directly as function parameters.
+    If both are provided, function parameters take precedence.
+    """
+    if isinstance(filename, Iterable) and not isinstance(filename, str):
+        return record_to_nested_dictionary(filename)
+
+    filename_str = str(filename)
+    if "::" in filename_str:
+        filename, info = filename_str.rsplit("::", maxsplit=1)
+        info = table_info_from_text(info)
+    else:
+        info = table_info_from_text("")
+    filename = Path(filename)
+    if sheet:
+        info.sheet = sheet
+    if key_column:
+        info.key_column = key_column
+    if value_column:
+        info.value_column = value_column
+
+    if str(filename).endswith(".json"):
+        return json.load(open(filename))
+    elif str(filename).endswith(".yaml") or str(filename).endswith(".yml"):
+        import yaml  # pylint: disable=import-outside-toplevel
+
+        if hasattr(yaml, "full_load"):
+            return yaml.full_load(open(filename))
+        return yaml.load(open(filename))  # pylint: disable=no-value-for-parameter
+    elif filename.suffix.lower() in [".csv", ".xlsx", ".ods"]:
+        return load_config_table(
+            filename,
+            sheet=info.sheet,
+            key_column=info.key_column,
+            value_column=info.value_column,
+        )
+    else:
+        sys.exit("Unknown file extension (file: {})".format(filename))
+
+
+def table_info_from_text(text, sheet=None, key_column=None, value_column=None):
+    """Convert comma-separated list of key-value pairs to info about table
+
+    Items are separated by comma. Key and value can be separated by
+    `=` or `:` and also by `: ` and ` = ` because spaces surrounding
+    key and value are removed.
+
+    If there is no key-value pair, the whole value is used as a
+    value for key_column.
+
+    Function parameters are used as default values.
+
+    Accepted keys are 'sheet', 'key_column', and 'value_column'.
+    """
+    info = types.SimpleNamespace(
+        sheet=sheet, key_column=key_column, value_column=value_column
+    )
+    if not text:
+        return info
+    items = text.split(",")
+    for item in items:
+        key = None
+        for separator in ["=", ":"]:
+            if separator in item:
+                key, value = item.split(separator, maxsplit=1)
+                key = key.strip()
+                value = value.strip()
+                break
+        if not key:
+            info.key_column = item
+        elif key in ["sheet", "key_column", "value_column"]:
+            setattr(info, key, value)
+        else:
+            raise ValueError(f"Unknown key '{key}' in table info specification")
+    return info
+
+
+def load_config_table(filename, sheet=None, key_column=None, value_column=None):
+    """Load a CSV file into a list of dictionaries
+
+    Values which can be converted into int or float are converted. Cells which can be
+    parsed as JSON, will be loaded into Python data structures (dicts, lists, etc.).
+
+    A whole file is read and loaded into memory unlike with the ``csv.reader()``
+    function.
+    """
+    if Path(filename).suffix.lower() not in [".csv", ".ods"]:
+        return load_config_xlsx(
+            filename, sheet=sheet, key_column=key_column, value_column=value_column
+        )
+
+    if Path(filename).suffix.lower() == ".ods":
+        return load_config_ods(
+            filename, sheet=sheet, key_column=key_column, value_column=value_column
+        )
+
+    return load_config_csv(filename, key_column=key_column, value_column=value_column)
+
+
+def load_config_csv(filename, key_column=None, value_column=None):
+    table = {}
+    # Read as CSV
+    with open(filename) as file:
+        # pylint: disable=import-outside-toplevel
+        import csv
+
+        try:
+            key_column = int(key_column)
+        except (TypeError, ValueError):
+            pass
+        try:
+            value_column = int(value_column)
+        except (TypeError, ValueError):
+            pass
+
+        if key_column is None:
+            key_column = 0
+        elif isinstance(key_column, str):
+            key_column = ord(key_column) - ord("A")
+        else:
+            key_column = int(key_column) - 1
+        if value_column is None:
+            value_column = key_column + 1
+        elif isinstance(value_column, str):
+            value_column = ord(value_column) - ord("A")
+        else:
+            value_column = int(value_column) - 1
+        for row in csv.reader(file):
+            key = validate_key(row[key_column])
+            if key:
+                value = text_to_value(row[value_column])
+                table[key] = value
+
+    return record_to_nested_dictionary(table)
+
+
+def validate_key(arg):
+    """Return a validated key or None if it cannot be validated.
+
+    Accepts any value which is supposed to be a configuration key
+    and returns that value ready to be used as a key, i.e., a string
+    without any extra whitespace around it (in case of user-input errors
+    in a spreadsheet).
+
+    For an empty string, NaN (not a number), and None returns None,
+    i.e., the key is invalid and the associated value in a key-value pair (if any)
+    should be ignored. Additionally, it returns None also for keys containing
+    a space (after being stripped from leading and trainling spaces). These are
+    assumed to be headings inside a spreadsheet or other values to be ignored
+    because valid keys don't have spaces in them.
+    """
+    if isinstance(arg, str):
+        # Allow for whitespace around the key.
+        arg = arg.strip()
+        if not arg or " " in arg:
+            return None
+        return arg
+    if arg is None:
+        return None
+    if isinstance(arg, float) and math.isnan(arg):
+        # NaN is None in our context since we don't use NaN in configuration
+        # and Pandas reader defaults to NaN for no-data values.
+        return None
+    # All keys should be strings, so convert numbers and anything else to strings.
+    return str(arg)
+
+
+def load_config_xlsx(filename, sheet=None, key_column=None, value_column=None):
+    """Load a CSV file into a list of dictionaries
+
+    Values which can be converted into int or float are converted. Cells which can be
+    parsed as JSON, will be loaded into Python data structures (dicts, lists, etc.).
+
+    A whole file is read and loaded into memory unlike with the ``csv.reader()``
+    function.
+    """
+    table = {}
+
+    # pylint: disable=import-outside-toplevel
+    import openpyxl
+    from openpyxl.utils import column_index_from_string
+
+    def column_from_string(arg, default):
+        if arg is None:
+            return default
+        else:
+            try:
+                return int(arg) - 1
+            except ValueError:
+                return column_index_from_string(arg) - 1
+
+    key_column = column_from_string(key_column, default=0)
+    value_column = column_from_string(value_column, default=key_column + 1)
+
+    import warnings
+
+    with warnings.catch_warnings():
+        # We want to use validation functions in the spreadsheet,
+        # but it does not matter whether they are supported by the reader here.
+        warnings.filterwarnings(
+            "ignore", message="Data Validation extension is not supported"
+        )
+        workbook = None
+        try:
+            workbook = openpyxl.load_workbook(filename, read_only=True)
+            if sheet:
+                sheet = workbook[sheet]
+            else:
+                sheet = workbook.active
+            # Read rows excluding the header.
+            for row in sheet.iter_rows(values_only=True):
+                key = validate_key(row[key_column])
+                if key:
+                    # Consider only rows with filled key to allow
+                    # for empty rows for formatting purposes.
+                    # Additionally, ignore rows where key cell contains
+                    # spaces (this allows for a header without threating
+                    # first row differently).
+                    value = text_to_value(row[value_column])
+                    table[key] = value
+        finally:
+            # Read-only mode requires an explicit close and
+            # the workbook object is not a context manager.
+            if workbook:
+                workbook.close()
+        return record_to_nested_dictionary(table)
+
+
+def load_config_ods(filename, sheet=None, key_column=None, value_column=None):
+    """Load a CSV file into a list of dictionaries
+
+    Values which can be converted into int or float are converted. Cells which can be
+    parsed as JSON, will be loaded into Python data structures (dicts, lists, etc.).
+
+    A whole file is read and loaded into memory unlike with the ``csv.reader()``
+    function.
+    """
+    table = {}
+
+    # pylint: disable=import-outside-toplevel
+    import pandas
+
+    sheet = sheet if sheet else 0
+
+    def column_from_string(arg, default):
+        if arg is None:
+            return default
+        else:
+            try:
+                return int(arg) - 1
+            except ValueError:
+                return arg
+
+    key_column = column_from_string(key_column, default=0)
+    value_column = column_from_string(value_column, default=1)
+
+    if isinstance(key_column, int) and isinstance(value_column, int):
+        cols = [key_column, value_column]
+        if key_column < value_column:
+            key_column_index = 1
+            value_column_index = 2
+        else:
+            key_column_index = 2
+            value_column_index = 1
+    else:
+        cols = f"{key_column},{value_column}"
+        # TODO: Defaults to flipped order for multi-letter column (flip it).
+        if (
+            len(key_column) == 1
+            and len(value_column) == 1
+            and ord(key_column) < ord(value_column)
+        ):
+            key_column_index = 1
+            value_column_index = 2
+        else:
+            key_column_index = 2
+            value_column_index = 1
+
+    data = pandas.read_excel(
+        Path(filename), sheet_name=sheet, header=None, usecols=cols
+    )
+    table = {}
+    for row in data.itertuples():
+        key = validate_key(row[key_column_index])
+        if key:
+            value = text_to_value(row[value_column_index])
+            if isinstance(value, float) and math.isnan(value):
+                value = None
+            table[key] = value
+    return record_to_nested_dictionary(table)
+
+
+def add_dict_config_to_table(table, value, keys=None):
+    if keys is None:
+        keys = []
+    if isinstance(value, Mapping):
+        for nested_key, nested_value in value.items():
+            new_keys = keys.copy()
+            new_keys.append(nested_key)
+            add_dict_config_to_table(table, value=nested_value, keys=new_keys)
+    else:
+        table["/".join(keys)] = value
+
+
+def dict_config_to_table(value):
+    table = {}
+    add_dict_config_to_table(table, value)
+    return table
+
+
+def print_table_config(config, file=None):
+    for key, value in config.items():
+        if isinstance(value, Iterable) and not isinstance(value, str):
+            value = json.dumps(value)
+        print(f"{key};{value}", file=file)
+
+
+def load_scenario_table(filename):
+    """Load a CSV file into a list of dictionaries
+
+    Values which can be converted into int or float are converted. Cells which can be
+    parsed as JSON, will be loaded into Python data structures (dicts, lists, etc.).
+
+    A whole file is read and loaded into memory unlike with the ``csv.reader()``
+    function.
+    """
+    # pylint: disable=import-outside-toplevel
+    table = []
+
+    # Read spreadsheet formats
+    if Path(filename).suffix.lower() != ".csv":
+        import openpyxl
+
+        try:
+            workbook = openpyxl.load_workbook(filename, read_only=True)
+            sheet = workbook.active
+            # Get header.
+            header = [cell.value for cell in sheet[1]]
+            # Read rows excluding the header.
+            for old_row in sheet.iter_rows(min_row=2):
+                new_row = {}
+                for key, cell in zip(header, old_row):
+                    new_row[key] = text_to_value(cell.value)
+                table.append(new_row)
+        finally:
+            # Read-only mode requires an explicit close and
+            # the workbook object is not a context manager.
+            workbook.close()
+        return table
+
+    # Read as CSV
+    with open(filename) as file:
+        import csv
+
+        for row in csv.DictReader(file):
+            for key, value in row.items():
+                row[key] = text_to_value(value)
+            table.append(row)
+    return table
