+++ popsborder/scenarios.py
@@ -20,73 +20,8 @@
 .. codeauthor:: Vaclav Petras <wenzeslaus gmail com>
 """
 
-import copy
-import collections.abc
-import json
-from pathlib import Path
-
 from .simulation import run_simulation
-
-
-def update_nested_dict_by_dict(dictionary, update):
-    """Recursively update nested dictionary by anther nested dictionary"""
-    for key, value in update.items():
-        if isinstance(value, collections.abc.Mapping):
-            update_nested_dict_by_dict(dictionary.get(key, {}), value)
-        else:
-            dictionary[key] = value
-
-
-def update_nested_dict_by_item(dictionary, keys, value):
-    """Update nested dictionary by a nested keys-value pair
-
-    An item is a list of keys to navigate the nested dictionary and a value to place
-    in the given position.
-
-    When a key can be represented as an int, it is used as a list index.
-    List must already exists in the given size or the only index used must be 0.
-
-    Floating point keys are not supported.
-    """
-    if len(keys) == 1:
-        key = keys[0]
-        try:
-            key = int(key)
-        except ValueError:
-            pass
-        dictionary[key] = value
-    else:
-        if keys[0] not in dictionary:
-            try:
-                # Test if the next key is an integer and thus index in a list.
-                int(keys[1])
-                # In case it is a list, we require keys are items are filled in order.
-                dictionary[keys[0]] = [None]
-            except ValueError:
-                dictionary[keys[0]] = {}
-        update_nested_dict_by_item(dictionary[keys[0]], keys[1:], value)
-
-
-def record_to_nested_dictionary(record):
-    """Convert dictionary with key/subkey/subsubkey keys into a nested dictionary"""
-    out = {}
-    for path, value in record.items():
-        try:
-            keys = path.split("/")
-        except AttributeError as error:
-            raise ValueError(
-                f"Record items need to be strings, not {path} ({type(path).__name__})"
-            ) from error
-        update_nested_dict_by_item(out, keys, value)
-    return out
-
-
-def update_config(config, record):
-    """Update config dictionary by a dictionary with key/subkey/subsubkey keys"""
-    config = copy.deepcopy(config)
-    update = record_to_nested_dictionary(record)
-    update_nested_dict_by_dict(config, update)
-    return config
+from .inputs import update_config
 
 
 def run_scenarios(
@@ -134,71 +69,3 @@ def run_scenarios(
             # The result is simulation totals.
             results.append((result, scenario_config))
     return results
-
-
-def text_to_value(arg):
-    """Convert text to int, float, or interpret it as JSON if possible
-
-    If the argument is not string, it is returned as is.
-    If conversion is not possible, the original text is returned.
-    None is returned for both None and an empty string.
-    """
-    if not isinstance(arg, str):
-        return arg
-    if not arg:
-        return None
-    try:
-        return int(arg)
-    except ValueError:
-        try:
-            return float(arg)
-        except ValueError:
-            try:
-                return json.loads(arg)
-            except json.JSONDecodeError:
-                # Return the original value.
-                return arg
-
-
-def load_scenario_table(filename):
-    """Load a CSV file into a list of dictionaries
-
-    Values which can be converted into int or float are converted. Cells which can be
-    parsed as JSON, will be loaded into Python data structures (dicts, lists, etc.).
-
-    A whole file is read and loaded into memory unlike with the ``csv.reader()``
-    function.
-    """
-    # pylint: disable=import-outside-toplevel
-    table = []
-
-    # Read spreadsheet formats
-    if Path(filename).suffix.lower() != ".csv":
-        import openpyxl
-
-        try:
-            workbook = openpyxl.load_workbook(filename, read_only=True)
-            sheet = workbook.active
-            # Get header.
-            header = [cell.value for cell in sheet[1]]
-            # Read rows excluding the header.
-            for old_row in sheet.iter_rows(min_row=2):
-                new_row = {}
-                for key, cell in zip(header, old_row):
-                    new_row[key] = text_to_value(cell.value)
-                table.append(new_row)
-        finally:
-            # Read-only mode requires an explicit close and
-            # the workbook object is not a context manager.
-            workbook.close()
-        return table
-
-    # Read as CSV
-    with open(filename) as file:
-        import csv
-
-        for row in csv.DictReader(file):
-            for key, value in row.items():
-                row[key] = text_to_value(value)
-            table.append(row)
-    return table
