+++ popsborder/simulation.py
@@ -22,11 +22,7 @@
 """
 
 import random
-import math
-import sys
 import types
-from collections.abc import Iterable, Mapping
-from pathlib import Path
 
 import numpy as np
 
@@ -379,298 +375,3 @@ def run_simulation(
         return details, totals
     else:
         return totals
-
-
-def load_configuration_yaml_from_text(text):
-    """Return configuration dictionary from YAML in a string"""
-    import yaml  # pylint: disable=import-outside-toplevel
-
-    if hasattr(yaml, "full_load"):
-        return yaml.full_load(text)
-    return yaml.load(text)  # pylint: disable=no-value-for-parameter
-
-
-def load_configuration(filename, sheet=None, key_column=None, value_column=None):
-    """Get the configuration from a JSON or YAML file
-
-    The format is decided based on the file extension.
-    It uses full_load() (FullLoader) to read YAML.
-
-    The parameter can be a string or a path object (path-like object).
-
-    If the *filename* contains `::`, anything after the last `::` is considered
-    parameters determining where in the spreadsheet or table are the relevant
-    columns. The format is multiple key-value pairs with key and value separated by
-    `=`, `:`, or `: ` and individual keys separated by `,`.
-    The same information can be passed directly as function parameters.
-    If both are provided, function parameters take precedence.
-    """
-    if isinstance(filename, Iterable) and not isinstance(filename, str):
-        from .scenarios import record_to_nested_dictionary
-
-        return record_to_nested_dictionary(filename)
-
-    filename_str = str(filename)
-    if "::" in filename_str:
-        filename, info = filename_str.rsplit("::", maxsplit=1)
-        info = table_info_from_text(info)
-    else:
-        info = table_info_from_text("")
-    filename = Path(filename)
-    if sheet:
-        info.sheet = sheet
-    if key_column:
-        info.key_column = key_column
-    if value_column:
-        info.value_column = value_column
-
-    if str(filename).endswith(".json"):
-        import json  # pylint: disable=import-outside-toplevel
-
-        return json.load(open(filename))
-    elif str(filename).endswith(".yaml") or str(filename).endswith(".yml"):
-        import yaml  # pylint: disable=import-outside-toplevel
-
-        if hasattr(yaml, "full_load"):
-            return yaml.full_load(open(filename))
-        return yaml.load(open(filename))  # pylint: disable=no-value-for-parameter
-    elif filename.suffix.lower() in [".csv", ".xlsx", ".ods"]:
-        return load_config_table(
-            filename,
-            sheet=info.sheet,
-            key_column=info.key_column,
-            value_column=info.value_column,
-        )
-    else:
-        sys.exit("Unknown file extension (file: {})".format(filename))
-
-
-def table_info_from_text(text, sheet=None, key_column=None, value_column=None):
-    """Convert comma-separated list of key-value pairs to info about table
-
-    Items are separated by comma. Key and value can be separated by
-    `=` or `:` and also by `: ` and ` = ` because spaces surrounding
-    key and value are removed.
-
-    If there is no key-value pair, the whole value is used as a
-    value for key_column.
-
-    Function parameters are used as default values.
-
-    Accepted keys are 'sheet', 'key_column', and 'value_column'.
-    """
-    info = types.SimpleNamespace(
-        sheet=sheet, key_column=key_column, value_column=value_column
-    )
-    if not text:
-        return info
-    items = text.split(",")
-    for item in items:
-        key = None
-        for separator in ["=", ":"]:
-            if separator in item:
-                key, value = item.split(separator, maxsplit=1)
-                key = key.strip()
-                value = value.strip()
-                break
-        if not key:
-            info.key_column = item
-        elif key in ["sheet", "key_column", "value_column"]:
-            setattr(info, key, value)
-        else:
-            raise ValueError(f"Unknown key '{key}' in table info specification")
-    return info
-
-
-def load_config_table(filename, sheet=None, key_column=None, value_column=None):
-    """Load a CSV file into a list of dictionaries
-
-    Values which can be converted into int or float are converted. Cells which can be
-    parsed as JSON, will be loaded into Python data structures (dicts, lists, etc.).
-
-    A whole file is read and loaded into memory unlike with the ``csv.reader()``
-    function.
-    """
-    # pylint: disable=import-outside-toplevel
-    table = {}
-
-    from .scenarios import record_to_nested_dictionary, text_to_value
-
-    # Read spreadsheet formats
-    if Path(filename).suffix.lower() not in [".csv", ".ods"]:
-        import openpyxl
-        from openpyxl.utils import column_index_from_string
-
-        if key_column is None:
-            key_column = 0
-        else:
-            try:
-                key_column = int(key_column) - 1
-            except ValueError:
-                key_column = column_index_from_string(key_column) - 1
-        if value_column is None:
-            value_column = key_column + 1
-        else:
-            try:
-                value_column = int(value_column) - 1
-            except ValueError:
-                value_column = column_index_from_string(value_column) - 1
-
-        import warnings
-
-        with warnings.catch_warnings():
-            # We want to use validation functions in the spreadsheet,
-            # but it does not matter whether they are supported by the reader here.
-            warnings.filterwarnings(
-                "ignore", message="Data Validation extension is not supported"
-            )
-            workbook = None
-            try:
-                workbook = openpyxl.load_workbook(filename, read_only=True)
-                if sheet:
-                    sheet = workbook[sheet]
-                else:
-                    sheet = workbook.active
-                # Read rows excluding the header.
-                for row in sheet.iter_rows(values_only=True):
-                    key = row[key_column]
-                    # All keys should be strings, so converst numbers to strings.
-                    # None needs to be empty string, not a string None.
-                    key = str(key) if key else ""
-                    # Allow for whitespace around the key.
-                    key = key.strip()
-                    if key and " " not in key:
-                        # Consider only rows with filled key to allow
-                        # for empty rows for formatting purposes.
-                        # Additionally, ignore rows where key cell contains
-                        # spaces (this allows for a header without threating
-                        # first row differently).
-                        value = text_to_value(row[value_column])
-                        table[key] = value
-            finally:
-                # Read-only mode requires an explicit close and
-                # the workbook object is not a context manager.
-                if workbook:
-                    workbook.close()
-            return record_to_nested_dictionary(table)
-
-    if Path(filename).suffix.lower() == ".ods":
-        import pandas
-
-        sheet = sheet if sheet else 0
-        if key_column is None:
-            key_column = 0
-        else:
-            try:
-                key_column = int(key_column) - 1
-            except ValueError:
-                pass
-        if value_column is None:
-            value_column = 1
-        else:
-            try:
-                value_column = int(value_column) - 1
-            except ValueError:
-                pass
-        if isinstance(key_column, int) and isinstance(value_column, int):
-            cols = [key_column, value_column]
-            if key_column < value_column:
-                key_column_index = 1
-                value_column_index = 2
-            else:
-                key_column_index = 2
-                value_column_index = 1
-        else:
-            cols = f"{key_column},{value_column}"
-            if (
-                len(key_column) == 1
-                and len(value_column) == 1
-                and ord(key_column) < ord(value_column)
-            ):
-                key_column_index = 1
-                value_column_index = 2
-            else:
-                key_column_index = 2
-                value_column_index = 1
-        data = pandas.read_excel(
-            Path(filename), sheet_name=sheet, header=None, usecols=cols
-        )
-        table = {}
-        for row in data.itertuples():
-            key = row[key_column_index]
-            if key and not (isinstance(key, float) and math.isnan(key)):
-                key = str(key)
-            else:
-                key = ""
-            key = key.strip()
-            if key and " " not in key:
-                value = text_to_value(row[value_column_index])
-                if isinstance(value, float) and math.isnan(value):
-                    value = None
-                table[key] = value
-        return record_to_nested_dictionary(table)
-
-    # Read as CSV
-    with open(filename) as file:
-        import csv
-
-        try:
-            key_column = int(key_column)
-        except (TypeError, ValueError):
-            pass
-        try:
-            value_column = int(value_column)
-        except (TypeError, ValueError):
-            pass
-
-        if key_column is None:
-            key_column = 0
-        elif isinstance(key_column, str):
-            key_column = ord(key_column) - ord("A")
-        else:
-            key_column = int(key_column) - 1
-        if value_column is None:
-            value_column = key_column + 1
-        elif isinstance(value_column, str):
-            value_column = ord(value_column) - ord("A")
-        else:
-            value_column = int(value_column) - 1
-        for row in csv.reader(file):
-            key = row[key_column]
-            # All keys should be strings, so converst numbers to strings.
-            # None needs to be empty string, not a string None.
-            key = str(key) if key else ""
-            # Allow for whitespace around the key.
-            key = key.strip()
-            if key and " " not in key:
-                value = text_to_value(row[value_column])
-                table[key] = value
-
-    return record_to_nested_dictionary(table)
-
-
-def add_dict_config_to_table(table, value, keys=None):
-    if keys is None:
-        keys = []
-    if isinstance(value, Mapping):
-        for nested_key, nested_value in value.items():
-            new_keys = keys.copy()
-            new_keys.append(nested_key)
-            add_dict_config_to_table(table, value=nested_value, keys=new_keys)
-    else:
-        table["/".join(keys)] = value
-
-
-def dict_config_to_table(value):
-    table = {}
-    add_dict_config_to_table(table, value)
-    return table
-
-
-def print_table_config(config, file=None):
-    for key, value in config.items():
-        if isinstance(value, Iterable) and not isinstance(value, str):
-            import json
-
-            value = json.dumps(value)
-        print(f"{key};{value}", file=file)
