+++ tests/test_run_simulation.py
@@ -1,5 +1,6 @@
 import pytest
-from popsborder.simulation import run_simulation, load_configuration_yaml_from_text
+from popsborder.inputs import load_configuration_yaml_from_text
+from popsborder.simulation import run_simulation
 
 
 CONFIG = """\
